+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-from-3/expected.js
@@ -1,0 +1,19 @@
+define(["exports", "foo"], function (exports, _foo) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  Object.defineProperty(exports, "foo", {
+    enumerable: true,
+    get: function () {
+      return _foo.foo;
+    }
+  });
+  Object.defineProperty(exports, "bar", {
+    enumerable: true,
+    get: function () {
+      return _foo.bar;
+    }
+  });
+});
