+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-from-5/actual.js
@@ -1,0 +1,1 @@
+export {foo as default} from "foo";
