+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-from-5/expected.js
@@ -1,0 +1,13 @@
+define(["exports", "foo"], function (exports, _foo) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  Object.defineProperty(exports, "default", {
+    enumerable: true,
+    get: function () {
+      return _foo.foo;
+    }
+  });
+});
