+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-from/expected.js
@@ -1,0 +1,16 @@
+define(["exports", "foo"], function (exports, _foo) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  Object.keys(_foo).forEach(function (key) {
+    if (key === "default" || key === "__esModule") return;
+    Object.defineProperty(exports, key, {
+      enumerable: true,
+      get: function () {
+        return _foo[key];
+      }
+    });
+  });
+});
