+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-named-2/expected.js
@@ -5,10 +5,5 @@ define(["exports"], function (exports) {
     value: true
   });
   exports.foo = foo;
-  exports.foo = foo;
-  exports.bar = bar;
-  exports.bar = foo;
-  exports.default = foo;
-  exports.default = foo;
   exports.bar = bar;
 });
