+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-named-5/expected.js
@@ -1,0 +1,9 @@
+define(["exports"], function (exports) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  exports.default = foo;
+  exports.bar = bar;
+});
