+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-named/actual.js
@@ -1,0 +1,1 @@
+export {foo};
