+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/export-named/expected.js
@@ -1,0 +1,8 @@
+define(["exports"], function (exports) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  exports.foo = foo;
+});
