--- packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/exports-from/expected.js
@@ -1,58 +1,0 @@
-define(["exports", "foo"], function (exports, _foo) {
-  "use strict";
-
-  Object.defineProperty(exports, "__esModule", {
-    value: true
-  });
-  Object.keys(_foo).forEach(function (key) {
-    if (key === "default" || key === "__esModule") return;
-    Object.defineProperty(exports, key, {
-      enumerable: true,
-      get: function () {
-        return _foo[key];
-      }
-    });
-  });
-  Object.defineProperty(exports, "foo", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "foo", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.bar;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "default", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "default", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.bar;
-    }
-  });
-});
