+++ packages/babel-plugin-transform-es2015-modules-amd/test/fixtures/amd/exports-variable/actual.js
@@ -1,9 +1,9 @@
 export var foo = 1;
-export var foo = 1, bar = 2;
-export var foo2 = function () {};
-export var foo3;
-export let foo4 = 2;
-export let foo5;
-export const foo6 = 3;
-export function foo7 () {}
-export class foo8 {}
+export var foo2 = 1, bar = 2;
+export var foo3 = function () {};
+export var foo4;
+export let foo5 = 2;
+export let foo6;
+export const foo7 = 3;
+export function foo8 () {}
+export class foo9 {}
